added test, changed errors

+++ Backend/src/models/device_model.py
@@ -1,41 +1,7 @@
 import datetime
-import os
 import uuid
 from dateutil import parser
-
-
-import pymongo
-
-##############################################
-# Init DB
-try:
-
-    if os.environ.get("DATABASE_IP") is not None:
-        db_location = os.environ["DATABASE_IP"] + ":27017"
-    else:
-        db_location = "localhost:27017"
-
-    print("Trying to connect to Device-DB located at: " + db_location)
-    # mongo = pymongo.MongoClient(os.environ["DATABASE_IP"], 27017)
-    mongo = pymongo.MongoClient(db_location, serverSelectionTimeoutMS=1000)
-    db = mongo.smts
-    print("Trying to connect to DB located at: " + db_location + "...Successful")
-    print(db.last_status)
-
-except pymongo.errors.ConnectionFailure as err:
-
-    if os.environ.get("DATABASE_IP") is not None:
-        db_location = os.environ["DATABASE_IP"] + ":27017"
-    else:
-        db_location = "localhost:2017"
-
-    print(
-        "Trying to connect to Device-DB located at: "
-        + db_location
-        + "...Failed...Err: "
-        + err
-    )
-    exit(1)
+from models.db import db
 
 
 class Device:
@@ -72,16 +38,17 @@ class Device:
         if coll.find_one({"name": device["name"]}):
             return None, ValueError("Device name already exists!")
 
-        # check if device name is available
+        # check if device number is available
         if coll.find_one({"devicePhoneNumber": device["devicePhoneNumber"]}):
             return None, ValueError("Device number already exists!")
 
-        # check if device name is available
+        # check if device imei is available
         if coll.find_one({"imei": device["imei"]}):
             return None, ValueError("Device imei already exists!")
 
         # Everything okey and new device can be inserted!
         result = coll.insert_one(device)
+
         if not result.acknowledged:
             return None, ValueError("New Device coulnd't be added to DB")
 
@@ -153,38 +120,44 @@ class Device:
         del device["locations"]
         del device["battery"]
 
-        x = coll.find_one({"imei": imei})
+        device_check = coll.find_one({"imei": imei})
 
         # Check if device with specified id, belongs to current user
-        if not x or x["owner"] != current_user:
-            return ValueError("Device not updateable")
-            # return None, ValueError("Device not updateable")
+        if not device_check or device_check["owner"] != current_user:
+            return None, ValueError("Not allowed")
 
-        device["owner"] = current_user
+        #device["owner"] = current_user
         # Try Updating
         cursor = coll.update_one({"imei": device["imei"]}, {"$set": device})
+
         if not cursor.acknowledged:
             # return None, ValueError("Couldn't update the device object")
-            return ValueError("Couldn't update the device object")
+            return None, ValueError("DB couldn't update the device object")
 
         # Get new updated Object
         updated_device = coll.find_one({"imei": device["imei"]})
+
         if updated_device is None:
             # return None, ValueError("Couldn't get the new device object")
-            return ValueError("Couldn't get the new device object")
+            return None, ValueError("Couldn't get the new device object")
 
-        return updated_device
+        return updated_device, None
 
-    def delete_device(self, imei: str):
+    def delete_device(self, imei: str, user_id: str):
         devices_coll = db["devices"]
 
         # Get and check device
         device = devices_coll.find_one({"imei": imei})
+
+        if device["owner"] != user_id:
+            return False, ValueError("Not allowed")
+
         if device is None:
             return False, ValueError("Couldn't find devices to be deleted.")
 
         # Delete device object now. When not succeeding restore data in user
         deleted = devices_coll.delete_one({"imei": imei})
+
         if not deleted.acknowledged:
             return False, ValueError("Couldn't delete device. restored data in User")
 
@@ -211,7 +184,7 @@ class Device:
             # return status inactive -> tracker is not sending anything the past maxDelta time.
             return "inactive", None
 
-    def get_device_locations(self, imei, start_time: datetime.datetime, end_time: datetime.datetime):
+    def get_device_locations(self, user_id, imei, start_time: datetime.datetime, end_time: datetime.datetime):
 
         computed_start = None
         computed_end = None
@@ -223,12 +196,15 @@ class Device:
 
         # Get/check device
         device = coll.find_one({"imei": imei})
+
         if device is None:
-            return None, ValueError("Deivce not found for getting any location(s)")
+            return None, ValueError("Device not found for getting any location(s)")
+        if device["owner"] != user_id:
+            return None, ValueError("Not allowed")
 
         # Get/check locations property
         locations = device.get("locations")
-        if locations is None or len(locations) == 0:
+        if len(locations) == 0:
             return None, ValueError("No locations in device found")
 
         # Compute all locations from start to now
@@ -273,13 +249,17 @@ class Device:
         # return locations
         return retLocations, None
 
-    def delete_locations(self, imei: str):
+    def delete_locations(self, imei: str, user_id: str):
 
         # Get/check device
         coll = db["devices"]
         device = coll.find_one({"imei": imei})
+
         if device is None:
             return False, ValueError("Couldn't retrieve device data from DB")
+
+        if device["owner"] != user_id:
+            return False, ValueError("Not allowed")
 
         # Update and check if it was successful
         updated = coll.update_one({"imei": imei}, {"$set": {"locations": []}})
